# Replace ЧебышёвSeries with PolynomialInЧебышёвBasis in the approximation code