--- a/benchmarks/approximation_benchmark.cpp
+++ b/benchmarks/approximation_benchmark.cpp
@@ -1,5 +1,6 @@
 // .\Release\x64\benchmarks.exe --benchmark_repetitions=3 --benchmark_filter=ApproximationBenchmark --benchmark_min_time=30  // NOLINT(whitespace/line_length)
 
+#include <memory>
 #include <random>
 #include <vector>