--- a/numerics/approximation.hpp
+++ b/numerics/approximation.hpp
@@ -1,10 +1,11 @@
 #pragma once
 
 #include <functional>
+#include <memory>
 #include <type_traits>
 #include <vector>
 
-#include "numerics/чебышёв_series.hpp"
+#include "numerics/polynomial_in_чебышёв_basis.hpp"
 #include "quantities/named_quantities.hpp"
 
 namespace principia {
@@ -12,7 +13,7 @@ namespace numerics {
 namespace _approximation {
 namespace internal {
 
-using namespace principia::numerics::_чебышёв_series;
+using namespace principia::numerics::_polynomial_in_чебышёв_basis;
 using namespace principia::quantities::_named_quantities;
 
 template<typename Argument, typename Function>
@@ -22,21 +23,22 @@ using Value = std::invoke_result_t<Function, Argument>;
 // further.  It is only called if the |error_estimate| is larger than the
 // |max_error| given to |AdaptiveЧебышёвPolynomialInterpolant|.
 template<typename Value, typename Argument>
-using SubdivisionPredicate =
-    std::function<bool(ЧебышёвSeries<Value, Argument> const& interpolant,
-                       Difference<Value> const& error_estimate)>;
+using SubdivisionPredicate = std::function<bool(
+    PolynomialInЧебышёвBasis<Value, Argument> const& interpolant,
+    Difference<Value> const& error_estimate)>;
 
 // A function that returns true iff construction of the interpolants must stop.
 template<typename Value, typename Argument>
-using TerminationPredicate =
-    std::function<bool(ЧебышёвSeries<Value, Argument> interpolant)>;
+using TerminationPredicate = std::function<bool(
+    std::unique_ptr<PolynomialInЧебышёвBasis<Value, Argument>> interpolant)>;
 
 // Returns a Чебышёв polynomial interpolant of f over
 // [lower_bound, upper_bound].  Stops if the absolute error is estimated to be
 // below |max_error| or if |max_degree| has been reached.  If |error_estimate|
 // is nonnull, it receives the estimate of the L∞ error.
 template<int max_degree, typename Argument, typename Function>
-ЧебышёвSeries<Value<Argument, Function>, Argument> ЧебышёвPolynomialInterpolant(
+std::unique_ptr<PolynomialInЧебышёвBasis<Value<Argument, Function>, Argument>>
+ЧебышёвPolynomialInterpolant(
     Function const& f,
     Argument const& lower_bound,
     Argument const& upper_bound,
@@ -47,7 +49,8 @@ template<int max_degree, typename Argument, typename Function>
 // together cover [lower_bound, upper_bound].  Subdivides the interval until the
 // error is below |max_error| or |subdivide| returns false.
 template<int max_degree, typename Argument, typename Function>
-std::vector<ЧебышёвSeries<Value<Argument, Function>, Argument>>
+std::vector<std::unique_ptr<
+    PolynomialInЧебышёвBasis<Value<Argument, Function>, Argument>>>
 AdaptiveЧебышёвPolynomialInterpolant(
     Function const& f,
     Argument const& lower_bound,