--- a/numerics/polynomial_in_чебышёв_basis.hpp
+++ b/numerics/polynomial_in_чебышёв_basis.hpp
@@ -8,6 +8,7 @@
 
 #include <memory>
 #include <optional>
+#include <string>
 
 #include "absl/container/btree_set.h"
 #include "base/not_null.hpp"