--- a/numerics/polynomial_in_чебышёв_basis_body.hpp
+++ b/numerics/polynomial_in_чебышёв_basis_body.hpp
@@ -2,6 +2,7 @@
 
 #include "numerics/polynomial_in_чебышёв_basis.hpp"
 
+#include <memory>
 #include <utility>
 
 #include "base/not_constructible.hpp"