--- a/numerics/чебышёв_lobatto_body.hpp
+++ b/numerics/чебышёв_lobatto_body.hpp
@@ -21,7 +21,7 @@ using namespace principia::quantities::_si;
 // k (values for even k should be obtained by calling the N/2 instantiation).
 template<int N>
 double ЧебышёвLobattoPoint(std::int64_t const k) {
-  static std::array<double, N> points = [](){
+  static std::array<double, N> points = []() {
     std::array<double, N> points;
     for (std::int64_t k = 1; k <= N; k += 2) {
       points[k] = Cos(π * k * Radian / N);