Also some optimization of the Чебышёв-Lobatto points.

Benchmarks before:
```
-------------------------------------------------------------------------------------------------------------------
Benchmark                                                                         Time             CPU   Iterations
-------------------------------------------------------------------------------------------------------------------
ApproximationBenchmark<4>/AdaptiveЧебышёвPolynomialInterpolant4     21721212 ns     21484375 ns           32
ApproximationBenchmark<8>/AdaptiveЧебышёвPolynomialInterpolant8     16757837 ns     16387195 ns           41
ApproximationBenchmark<16>/AdaptiveЧебышёвPolynomialInterpolant16   17460602 ns     17708333 ns           45
ApproximationBenchmark<32>/AdaptiveЧебышёвPolynomialInterpolant32   20746094 ns     20680147 ns           34
```
after:
```
-------------------------------------------------------------------------------------------------------------------
Benchmark                                                                         Time             CPU   Iterations
-------------------------------------------------------------------------------------------------------------------
ApproximationBenchmark<4>/AdaptiveЧебышёвPolynomialInterpolant4     16597847 ns     16666667 ns           45
ApproximationBenchmark<8>/AdaptiveЧебышёвPolynomialInterpolant8     13742832 ns     13750000 ns           50
ApproximationBenchmark<16>/AdaptiveЧебышёвPolynomialInterpolant16   14725060 ns     14687500 ns           50
ApproximationBenchmark<32>/AdaptiveЧебышёвPolynomialInterpolant32   18054937 ns     18292683 ns           41
```